chore(extensions): move pages directory
